airflow/requirements.txt
apache-airflow==2.8.1
sqlalchemy==1.4.28
psycopg2-binary
apache-airflow-providers-docker
